updated app to use new react router
