+++ react-app/.env.development
@@ -2,7 +2,7 @@
 
 # Server namespace
 REACT_APP_HOST_URI=http://localhost:4502
-REACT_APP_GRAPHQL_ENDPOINT=/content/_cq_graphql/wknd/endpoint.json
+REACT_APP_GRAPHQL_ENDPOINT=/content/_cq_graphql/wknd-shared/endpoint.json
 
 #AUTH (Choose one method)
 # Authentication methods: 'service-token', 'dev-token', 'basic' or leave blank to use no authentication
