+++ react-app/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@adobe/aem-headless-client-js": "^3.1.0",
     "@adobe/aem-headless-client-nodejs": "^1.0.2",
+    "@babel/core": "^7.18.2",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
@@ -13,9 +14,10 @@
     "normalize-scss": "^7.0.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "3.4.3",
-    "sass": "^1.42.1"
+    "sass": "^1.42.1",
+    "typescript": "^4.7.3"
   },
   "scripts": {
     "start": "react-scripts start",
