+++ react-app/src/components/AdventureDetail.scss
@@ -18,6 +18,8 @@ it.
     width: 24px;
     float: right;
     margin: 1em;
+    background: none;
+    border: none;
   }
   
   .adventure-detail-title {
